EVP_AEAD: fixup cffi defs, add to _conditional

evp_aead: fix initialization on BoringSSL


_conditional: fatfingering


backends/openssl: make an AEAD helper private


openssl: collapse aead module


openssl/aead: experimenting


openssl/aead: use Cryptography_HAS_EVP_AEAD


openssl/aead: group things


Revert "openssl: collapse aead module"

This reverts commit 558b8d57f469b6abf51e70e68f5cf330b7ae6414.

aead: tweak feature test


evp_aead: stupidness


Revert "aead: tweak feature test"

This reverts commit aa2eea648fed3e7460cfa358bf89c646e56f310d.

Revert "Revert "openssl: collapse aead module""

This reverts commit 44a68c48b52dc0d243d23016999e7fafd580e48b.

Revert "Revert "Revert "openssl: collapse aead module"""

This reverts commit c35bb37f6c9b4a56f88d36402c8e13ab6396156b.

Revert "Revert "aead: tweak feature test""

This reverts commit 78c0dc5ed298c0fde236b8137cb2b0795e905edc.

openssl/aead: try to migrate more incrementally


aead: lintage


openssl/aead: more incremental rewriting


openssl/aead: rename _OpenSSL -> _EVPCIPHER


openssl/aead: collapse module
