+++ src/cryptography/hazmat/bindings/openssl/_conditional.py
@@ -125,7 +125,6 @@ def cryptography_has_tlsv13_functions() -> typing.List[str]:
         "SSL_set_post_handshake_auth",
         "SSL_SESSION_get_max_early_data",
         "SSL_write_early_data",
-        "SSL_read_early_data",
         "SSL_CTX_set_max_early_data",
     ]
 
@@ -262,6 +261,17 @@ def cryptography_has_evp_pkey_set_peer_ex() -> typing.List[str]:
     return ["EVP_PKEY_derive_set_peer_ex"]
 
 
+def cryptography_has_evp_aead() -> typing.List[str]:
+    return [
+        "EVP_aead_chacha20_poly1305",
+        "EVP_AEAD_CTX_free",
+        "EVP_AEAD_CTX_seal",
+        "EVP_AEAD_CTX_open",
+        "EVP_AEAD_max_overhead",
+        "Cryptography_EVP_AEAD_CTX_new",
+    ]
+
+
 # This is a mapping of
 # {condition: function-returning-names-dependent-on-that-condition} so we can
 # loop over them and delete unsupported names at runtime. It will be removed
@@ -314,4 +324,5 @@ CONDITIONAL_NAMES = {
     "Cryptography_HAS_EVP_PKEY_SET_PEER_EX": (
         cryptography_has_evp_pkey_set_peer_ex
     ),
+    "Cryptography_HAS_EVP_AEAD": (cryptography_has_evp_aead),
 }
